requirements.txt
requests
connexion[swagger-ui]
flask
coloredlogs
flask-cors
google-cloud-storage